IndexedDB/tests/submissions/Opera/idbindex_get-tests.htm

<!DOCTYPE html>
<meta charset="utf-8">
<title>IDBIndex.get()</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support.js"></script>

<script type="text/javascript">
//XXX Bah, not sure if this really is good for tests with idb API. No really good test fit
    var records = [ { key:1, value:"data" },
                    { key:2, value:"data" },
                    { key:3, value:"data" } ];

    function test_me(desc, func) {
        var t = async_test(desc);

        var open_rq = createdb(t);
        open_rq.onupgradeneeded = function(e) {
            var objStore = e.target.result.createObjectStore("test", { keyPath: "key" });
            objStore.createIndex("index", "indexedProperty");

            for (var i = 0; i < records.length; i++)
                objStore.add(records[i]);
        };

        open_rq.onsuccess = func
    }

    // ------- Tests starting -------

    test_me("returns the record",
        function(e) {
            var db = e.target.result,
              rq = db.transaction("test")
                     .objectStore("test")
                     .index("index")
                     .get("data");

            rq.onsuccess = this.step_func(function(e) {
                var record = e.target.result;

                assert_equals(record.key,   records[0].key);
                assert_equals(record.value, records[0].value);

                this.done();
            });
        });


    test_me("returns the record where the index contains duplicate values")
</script>

<div id="log"></div>
